Rename schema folder to definitions
