+++ src/definitions/edges.ts
