+++ src/definitions/variant.ts
